tes(addSurface): extract test factories
