+++ packages/support/lib-tools/src/project/test-utils/resourceFactory.ts
@@ -1,0 +1,6 @@
+import { UnknownResource } from '@noodles-ui/support-types';
+
+type ResourceType = 'surface' | 'theme';
+export function resourceFactory<T>(type: ResourceType, overides?: Partial<UnknownResource>): T {
+    return { type, name: '', module: '', extend: [], ...overides } as T;
+}
